Rename pyro.sugar.named -> pyro.contrib.named
