+++ pyro/contrib/named.py
